+++ backend/src/Controller/User/CharacterController.php
@@ -21,6 +21,7 @@ use OpenApi\Attributes as OA;
 use Psr\Http\Message\ResponseInterface;
 use Psr\Http\Message\ServerRequestInterface;
 use Psr\Log\LoggerInterface;
+use Swagger\Client\Eve\ApiException;
 use Swagger\Client\Eve\Model\GetCharactersCharacterIdNotFound;
 use Swagger\Client\Eve\Model\GetCharactersCharacterIdOk;
 
@@ -321,10 +322,21 @@ class CharacterController extends BaseController
             $eveChar = $esiApiFactory
                 ->getCharacterApi()
                 ->getCharactersCharacterId($charId, $config['eve']['datasource']);
+        } catch (ApiException $e) {
+            $body = $e->getResponseBody();
+            if ($e->getCode() === 404 && is_string($body) && str_contains($body, 'Character not found')) {
+                return $this->withJson('Character not found.', 404);
+            } else {
+                $log->error($e->getMessage());
+                return $this->withJson('ESI error.', 500);
+            }
         } catch (\Exception $e) {
             $log->error($e->getMessage());
             return $this->withJson('ESI error.', 500);
         }
+
+        // This is currently (tkhamez/swagger-eve-php v10.1.0) not used, but I think that's a bug, see
+        // https://github.com/OpenAPITools/openapi-generator/pull/19483
         if (
             $eveChar instanceof GetCharactersCharacterIdNotFound &&
             str_contains($eveChar->getError(), 'Character not found')