+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 dd mm yyyy
 
+- Fix: Deleted characters were no longer removed.
 - Changed default error reporting level to "E_ALL & ~E_DEPRECATED".